Refactor Description
