+++ src/main/java/seedu/address/model/common/Description.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.common;
 
 import static seedu.address.commons.util.AppUtil.checkArgument;
