+++ src/main/java/seedu/address/model/event/Event.java
@@ -11,7 +11,7 @@ import java.util.Set;
 
 import seedu.address.model.AddressBook;
 import seedu.address.model.person.Address;
-import seedu.address.model.person.Description;
+import seedu.address.model.common.Description;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.Name;
 import seedu.address.model.person.Person;
