+++ src/main/java/seedu/address/model/person/Person.java
@@ -9,6 +9,7 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 
+import seedu.address.model.common.Description;
 import seedu.address.model.tag.Tag;
 
 /**
